+++ tests/models/pony.py
@@ -18,7 +18,7 @@ db = Database(databases[current].pop('type'), *databases[current]['args'], **dat
 # Generation of entities for date range partitioning
 for item in ('day', 'week', 'month', 'year'):
     name = 'RangeDate{0}'.format(item.capitalize())
-    partition = install('partition', type='range', subtype='date', range=item, column='created')
+    partition = install('partition', type='range', subtype='date', constraint=item, column='created')
 
     locals()[name] = partition(type(name, (db.Entity,), {
         '_table_': 'test_rangedate{0}'.format(item),
@@ -30,7 +30,7 @@ if not os.environ.get('DB') == 'mysql':
     # Generation of entities for integer range partitioning
     for item in ('2', '5'):
         name = 'RangeInteger{0}'.format(item)
-        partition = install('partition', type='range', subtype='integer', range=item, column='num')
+        partition = install('partition', type='range', subtype='integer', constraint=item, column='num')
 
         locals()[name] = partition(type(name, (db.Entity,), {
             '_table_': 'test_rangeinteger{0}'.format(item),
@@ -42,7 +42,7 @@ if not os.environ.get('DB') == 'mysql':
     for subtype in ('string_firstchars', 'string_lastchars'):
         for item in ('2', '5'):
             name = 'Range{0}{1}'.format(''.join(s.capitalize() for s in subtype.split('_')), item)
-            partition = install('partition', type='range', subtype=subtype, range=item, column='title')
+            partition = install('partition', type='range', subtype=subtype, constraint=item, column='title')
 
             locals()[name] = partition(type(name, (db.Entity,), {
                 '_table_': 'test_range{0}{1}'.format(subtype, item),