+++ tests/models/sqlalchemy.py
@@ -20,7 +20,7 @@ Base = declarative_base()
 # Generation of entities for date range partitioning
 for item in ('day', 'week', 'month', 'year'):
     name = 'RangeDate{0}'.format(item.capitalize())
-    partition = install('partition', type='range', subtype='date', range=item, column='created', dsn=engine.url)
+    partition = install('partition', type='range', subtype='date', constraint=item, column='created', dsn=engine.url)
 
     locals()[name] = partition(type(name, (Base,), {
         '__tablename__': 'test_rangedate{0}'.format(item),
@@ -33,7 +33,7 @@ if not os.environ.get('DB') == 'mysql':
     # Generation of entities for integer range partitioning
     for item in ('2', '5'):
         name = 'RangeInteger{0}'.format(item)
-        partition = install('partition', type='range', subtype='integer', range=item, column='num', dsn=engine.url)
+        partition = install('partition', type='range', subtype='integer', constraint=item, column='num', dsn=engine.url)
 
         locals()[name] = partition(type(name, (Base,), {
             '__tablename__': 'test_rangeinteger{0}'.format(item),
@@ -46,7 +46,7 @@ if not os.environ.get('DB') == 'mysql':
     for subtype in ('string_firstchars', 'string_lastchars'):
         for item in ('2', '5'):
             name = 'Range{0}{1}'.format(''.join(s.capitalize() for s in subtype.split('_')), item)
-            partition = install('partition', type='range', subtype=subtype, range=item, column='title', dsn=engine.url)
+            partition = install('partition', type='range', subtype=subtype, constraint=item, column='title', dsn=engine.url)
 
             locals()[name] = partition(type(name, (Base,), {
                 '__tablename__': 'test_range{0}{1}'.format(subtype, item),